refactor: restructure into explicit packages
- Dissolve internal/secrets/ into focused sub-packages:
  - internal/providers/ — Provider interface + Registry
  - internal/providers/bitwarden/ — BWClient, Cache, BWRef
  - internal/providers/vault/ — VaultClient, VaultRef
  - internal/state/ — shell state (var-name tracking, unload sentinel)

- Add internal/kubeconfig/source_uri.go (NormalizeSourceURI)
  to eliminate URI normalisation duplication between kctx and envoke

- Split large CLI files into per-command files:
  - internal/cli/renv/: resolve, exec, yaml_cmd, shell, status, cache, unload, watch, registry
  - internal/cli/kctx/: load, switch, shell, status, cache, unload, watch, registry
  - cmd/envoke/: resolve, unload, shell, cache_watch per-command files

- cmd/envoke/main.go reduced to main(), rootCmd(), newRegistry(),
  normalizeKubeconfigURI() — all command logic lives in dedicated files

- internal/env updated to accept *providers.Registry instead of
  concrete BWClient/VaultClient types
