+++ internal/providers/bitwarden/client.go
@@ -1,4 +1,4 @@
-package secrets
+package bitwarden
 
 import (
 	"bytes"
