+++ internal/providers/bitwarden/ref.go
@@ -1,4 +1,4 @@
-package secrets
+package bitwarden
 
 import (
 	"fmt"
@@ -11,12 +11,6 @@ type BWRef struct {
 	Item         string
 	FieldSpec    string // "password" (default), "username", "note", "totp", "field:<name>"
 	IsCollection bool   // true if bw://collection:<name>/item
-}
-
-// VaultRef holds the parsed components of a vault:// URI.
-type VaultRef struct {
-	Path  string
-	Field string
 }
 
 // ParseBWRef parses a bw://folder/item-name[/field-spec] URI.
@@ -50,31 +44,4 @@ func ParseBWRef(uri string) (BWRef, error) {
 		fieldSpec = parts[2]
 	}
 	return BWRef{Folder: folder, Item: item, FieldSpec: fieldSpec, IsCollection: isCollection}, nil
-}
-
-// ParseVaultRef parses a vault://path#field URI.
-// Returns an error if the field fragment is missing.
-func ParseVaultRef(uri string) (VaultRef, error) {
-	if !strings.HasPrefix(uri, "vault://") {
-		return VaultRef{}, fmt.Errorf("not a vault:// URI: %q", uri)
-	}
-	rest := strings.TrimPrefix(uri, "vault://")
-	idx := strings.LastIndex(rest, "#")
-	if idx < 0 {
-		return VaultRef{}, fmt.Errorf("vault:// URI missing #field fragment: %q", uri)
-	}
-	path := rest[:idx]
-	field := rest[idx+1:]
-	if path == "" {
-		return VaultRef{}, fmt.Errorf("vault:// URI has empty path: %q", uri)
-	}
-	if field == "" {
-		return VaultRef{}, fmt.Errorf("vault:// URI has empty field: %q", uri)
-	}
-	return VaultRef{Path: path, Field: field}, nil
-}
-
-// IsSecretRef returns true if s starts with "bw://" or "vault://".
-func IsSecretRef(s string) bool {
-	return strings.HasPrefix(s, "bw://") || strings.HasPrefix(s, "vault://")
 }
