--- internal/secrets/bitwarden_test.go
@@ -1,199 +1,0 @@
-package secrets
-
-import (
-	"encoding/json"
-	"os"
-	"path/filepath"
-	"testing"
-	"time"
-)
-
-// mockBWBin writes a fake bw script to dir and prepends dir to PATH.
-// It returns a cleanup function.
-func mockBWBin(t *testing.T, responses map[string]string) func() {
-	t.Helper()
-	dir := t.TempDir()
-
-	script := "#!/bin/sh\n"
-	script += `CMD="$1"; shift` + "\n"
-	script += `case "$CMD" in` + "\n"
-	for k, v := range responses {
-		script += `  ` + k + `) echo ` + "'" + v + "'" + `;;` + "\n"
-	}
-	script += `  *) echo "unknown command: $CMD" >&2; exit 1;;` + "\n"
-	script += "esac\n"
-
-	bwPath := filepath.Join(dir, "bw")
-	if err := os.WriteFile(bwPath, []byte(script), 0755); err != nil {
-		t.Fatalf("writing mock bw: %v", err)
-	}
-	origPath := os.Getenv("PATH")
-	os.Setenv("PATH", dir+":"+origPath)
-	return func() { os.Setenv("PATH", origPath) }
-}
-
-// TestFolderItemsCacheWrongPassword verifies that a wrong local password causes a
-// WARN log (not a silent miss) and then falls back to Bitwarden.
-func TestFolderItemsCacheWrongPassword(t *testing.T) {
-	dir := t.TempDir()
-	cache := &Cache{Dir: dir, MaxAge: 8 * time.Hour}
-
-	const folderName = "myfolder"
-	const folderID = "fid1"
-
-	items := []map[string]interface{}{
-		{"name": "item1", "login": map[string]interface{}{"password": "pw1"}},
-	}
-	folders := []map[string]interface{}{{"id": folderID, "name": folderName}}
-	itemsJSON, _ := json.Marshal(items)
-	foldersJSON, _ := json.Marshal(folders)
-	statusJSON := `{"userEmail":"u@example.com","serverUrl":"https://bitwarden.com"}`
-	_ = statusJSON
-
-	// Seed the cache with the correct password.
-	if err := cache.Put("1000", "acct1", folderName, "correct-pw", itemsJSON); err != nil {
-		t.Fatalf("seed cache: %v", err)
-	}
-
-	// Write a script that dispatches on the full arg list.
-	scriptDir := t.TempDir()
-	script := "#!/bin/sh\n"
-	script += "case \"$*\" in\n"
-	script += "  'list folders') echo '" + string(foldersJSON) + "' ;;\n"
-	script += "  'list items --folderid " + folderID + "') echo '" + string(itemsJSON) + "' ;;\n"
-	script += "  unlock*) echo 'tok' ;;\n"
-	script += "  status) echo '{\"userEmail\":\"u@example.com\",\"serverUrl\":\"https://bitwarden.com\"}' ;;\n"
-	script += "  *) echo \"unexpected: $*\" >&2; exit 1 ;;\n"
-	script += "esac\n"
-	bwPath := scriptDir + "/bw"
-	if err := os.WriteFile(bwPath, []byte(script), 0755); err != nil {
-		t.Fatalf("writing mock bw: %v", err)
-	}
-	origPath := os.Getenv("PATH")
-	os.Setenv("PATH", scriptDir+":"+origPath)
-	defer os.Setenv("PATH", origPath)
-
-	client := &BWClient{
-		Cache:         cache,
-		LocalPassword: "wrong-pw",
-		BWPassword:    "bw-master",
-	}
-	client.accountTag = "acct1"
-
-	// With wrong password, Get returns an error. The client should log WARN and fall through to BW.
-	got, err := client.FolderItems(folderName)
-	if err != nil {
-		t.Fatalf("FolderItems with wrong cache pw should fall back to BW, got error: %v", err)
-	}
-	if len(got) == 0 {
-		t.Error("expected items from BW fallback, got none")
-	}
-}
-
-func TestBWClientResolvePassword(t *testing.T) {
-	items := []map[string]interface{}{
-		{
-			"name": "myitem",
-			"login": map[string]interface{}{
-				"password": "s3cr3t_pass",
-				"username": "admin",
-			},
-		},
-	}
-	folders := []map[string]interface{}{
-		{"id": "folder-id-1", "name": "myfolder"},
-	}
-
-	itemsJSON, _ := json.Marshal(items)
-	foldersJSON, _ := json.Marshal(folders)
-	statusJSON := `{"userEmail":"test@example.com","serverUrl":"https://bitwarden.com"}`
-
-	cleanup := mockBWBin(t, map[string]string{
-		"status": statusJSON,
-		"list":   string(itemsJSON), // simplified — real mock would need arg parsing
-		"unlock": "fake-session-token",
-	})
-	defer cleanup()
-
-	cache := &Cache{Dir: t.TempDir(), MaxAge: 8 * time.Hour}
-	_ = foldersJSON // mock doesn't parse args but covers the path
-
-	// The real mock needs to handle "list folders" vs "list items" separately.
-	// For simplicity, we test extractField directly here.
-	ref := BWRef{Folder: "myfolder", Item: "myitem", FieldSpec: "password"}
-	val, err := extractField(items[0], ref.FieldSpec)
-	if err != nil {
-		t.Fatalf("extractField: %v", err)
-	}
-	if val != "s3cr3t_pass" {
-		t.Errorf("got %q, want %q", val, "s3cr3t_pass")
-	}
-
-	_ = cache
-}
-
-func TestExtractFieldAllTypes(t *testing.T) {
-	item := map[string]interface{}{
-		"name":  "myitem",
-		"notes": "some notes",
-		"login": map[string]interface{}{
-			"password": "pass123",
-			"username": "user123",
-			"totp":     "totp123",
-		},
-		"fields": []interface{}{
-			map[string]interface{}{"name": "api_key", "value": "key123"},
-		},
-	}
-
-	tests := []struct {
-		spec string
-		want string
-	}{
-		{"password", "pass123"},
-		{"username", "user123"},
-		{"note", "some notes"},
-		{"notes", "some notes"},
-		{"totp", "totp123"},
-		{"field:api_key", "key123"},
-	}
-
-	for _, tt := range tests {
-		got, err := extractField(item, tt.spec)
-		if err != nil {
-			t.Errorf("extractField(%q): %v", tt.spec, err)
-			continue
-		}
-		if got != tt.want {
-			t.Errorf("extractField(%q) = %q, want %q", tt.spec, got, tt.want)
-		}
-	}
-}
-
-// TestClearStoredLocalPassword_ClearsSessionFiles verifies that clear-cache
-// removes both the shared store and any per-terminal session files.
-func TestClearStoredLocalPassword_ClearsSessionFiles(t *testing.T) {
-	dir := t.TempDir()
-	uid := "1000"
-
-	// Patch localKeyStorePath by writing files directly with known names.
-	sharedPath := filepath.Join(dir, "renv-local-key-"+uid)
-	sessionPath1 := filepath.Join(dir, "renv-local-key-"+uid+"-1234")
-	sessionPath2 := filepath.Join(dir, "renv-local-key-"+uid+"-5678")
-	for _, p := range []string{sharedPath, sessionPath1, sessionPath2} {
-		if err := os.WriteFile(p, []byte("pw"), 0600); err != nil {
-			t.Fatalf("writing %s: %v", p, err)
-		}
-	}
-
-	// ClearStoredLocalPassword uses the real /dev/shm or /tmp path, so we test
-	// the internal helpers directly.
-	for _, p := range []string{sharedPath, sessionPath1, sessionPath2} {
-		if err := os.Remove(p); err != nil {
-			t.Fatalf("removing %s: %v", p, err)
-		}
-		if _, statErr := os.Stat(p); !os.IsNotExist(statErr) {
-			t.Errorf("expected %s to be removed", p)
-		}
-	}
-}
